fix/error-protection

+++ src/admin/controllers/user/controllerUser.js
@@ -15,6 +15,12 @@ module.exports = {
         try {
             const { id } = req.params;
 
+            const checkUser = await models.User.findByPk(id);
+
+            if (!checkUser) {
+                return res.status(400).json("This user does not exist!");
+            }
+
             const searchingUser = await models.User.findAll({
                 where: {
                     id
@@ -27,7 +33,7 @@ module.exports = {
         }
     },
     async addingUser(req, res) {
-        try{
+        try {
             const {
                 username, password
             } = req.body;
@@ -39,16 +45,26 @@ module.exports = {
             });
 
             res.status(201).json("User registered successfully!");
-        }catch(error){
-            res.status(500).json("An unexpected error has occurred!");
+        } catch (error) {
+            if (error.name == "SequelizeUniqueConstraintError") {
+                return res.status(403).json("This user is already being used!");
+            }
+
+            res.status(500).json("An unexpected error has occurred!" + error);
         }
     },
     async updatedUser(req, res) {
-        try{
+        try {
             const { id } = req.params;
-            const { 
+            const {
                 username, password
             } = req.body;
+
+            const checkUser = await models.User.findByPk(id);
+
+            if (!checkUser) {
+                return res.status(400).json("This user does not exist!");
+            }
 
             const new_password = bcryptjs.hashSync(password, 10);
 
@@ -61,13 +77,23 @@ module.exports = {
             });
 
             res.status(201).json("User successfully updated!");
-        }catch(error){
-            res.status(500).json("An unexpected error has occurred!");
+        } catch (error) {
+            if (error.name == "SequelizeUniqueConstraintError") {
+                return res.status(403).json("This user is already being used!");
+            }
+
+            res.status(500).json("An unexpected error has occurred!" + error);
         }
     },
     async removeUser(req, res) {
-        try{
+        try {
             const { id } = req.params;
+
+            const checkUser = await models.User.findByPk(id);
+
+            if (!checkUser) {
+                return res.status(400).json("This user does not exist!");
+            }
 
             await models.User.destroy({
                 where: {
@@ -76,7 +102,7 @@ module.exports = {
             });
 
             res.status(200).json("User successfully removed!");
-        }catch(error){
+        } catch (error) {
             res.status(500).json("An unexpected error has occurred!");
         }
     }
